+++ druid-lookups/src/main/java/com/yahoo/maha/maha_druid_lookups/query/lookup/namespace/URIExtractionNamespace.java
@@ -33,6 +33,7 @@ import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 
+import com.yahoo.maha.maha_druid_lookups.query.lookup.parser.FlatDataParser;
 import com.yahoo.maha.maha_druid_lookups.query.lookup.util.LookupUtil;
 import org.apache.druid.java.util.common.IAE;
 
@@ -235,179 +236,6 @@ public class URIExtractionNamespace implements OnlineDatastoreExtractionNamespac
         result = 31 * result * Objects.hashCode(maxHeapPercentage);
         return result;
     }
-
-    private static class DelegateParser implements Parser<String, List<String>>
-    {
-        private final Parser<String, Object> delegate;
-        private final String key;
-        private final String nullReplacement;
-
-        private DelegateParser(
-                Parser<String, Object> delegate,
-                String key,
-                String nullReplacement
-        )
-        {
-            this.delegate = delegate;
-            this.key = key;
-            this.nullReplacement = nullReplacement;
-        }
-
-        @Override
-        public Map<String, List<String>> parseToMap(String input)
-        {
-            final Map<String, Object> inner = delegate.parseToMap(input);
-            if (null == inner) {
-                // Skip null or missing values, treat them as if there were no row at all.
-                return ImmutableMap.of();
-            }
-
-            final String k = inner.get(key) == null ? nullReplacement : inner.get(key).toString();
-
-            return ImmutableMap.of(k, new ArrayList(inner.values()));
-        }
-
-        @Override
-        public void setFieldNames(Iterable<String> fieldNames)
-        {
-            delegate.setFieldNames(fieldNames);
-        }
-
-        @Override
-        public List<String> getFieldNames()
-        {
-            return delegate.getFieldNames();
-        }
-    }
-
-    @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "format")
-    @JsonSubTypes(value = {
-            @JsonSubTypes.Type(name = "csv", value = CSVFlatDataParser.class)
-    })
-    public interface FlatDataParser
-    {
-        Parser<String, List<String>> getParser();
-        List<String> getColumns();
-        String getKey();
-        String getNullReplacement();
-    }
-
-    @JsonTypeName("csv")
-    public static class CSVFlatDataParser implements FlatDataParser
-    {
-        private final Parser<String, List<String>> parser;
-        private final List<String> columns;
-        private final String keyColumn;
-        private final String nullReplacement;
-
-        @JsonCreator
-        public CSVFlatDataParser(
-                @JsonProperty("columns") List<String> columns,
-                @JsonProperty("keyColumn") final String keyColumn,
-                @JsonProperty("hasHeaderRow") boolean hasHeaderRow,
-                @JsonProperty("skipHeaderRows") int skipHeaderRows,
-                @JsonProperty("nullReplacement") String nullReplacement
-        )
-        {
-            Preconditions.checkArgument(
-                    Preconditions.checkNotNull(columns, "`columns` list required").size() > 1,
-                    "Must specify more than one column to have a key value pair"
-            );
-
-            this.columns = columns;
-            this.keyColumn = Strings.isNullOrEmpty(keyColumn) ? columns.get(0) : keyColumn;
-            this.nullReplacement = nullReplacement == null ? "" : nullReplacement;
-            Preconditions.checkArgument(
-                    columns.contains(this.keyColumn),
-                    "Column [%s] not found int columns: %s",
-                    this.keyColumn,
-                    Arrays.toString(columns.toArray())
-            );
-            CSVParser csvParser = new CSVParser(null, columns, hasHeaderRow, skipHeaderRows);
-            csvParser.startFileFromBeginning();
-            this.parser = new DelegateParser(
-                    csvParser,
-                    this.keyColumn,
-                    this.nullReplacement
-            );
-        }
-
-        @JsonProperty
-        public List<String> getColumns() { return columns; }
-
-        @JsonProperty
-        public String getKey() { return keyColumn; }
-
-        @JsonProperty
-        public String getNullReplacement() { return nullReplacement; }
-
-        @VisibleForTesting
-        CSVFlatDataParser(
-                List<String> columns,
-                String keyColumn,
-                String nullReplacement
-        )
-        {
-            this(columns, keyColumn, false, 0, nullReplacement);
-        }
-
-        @VisibleForTesting
-        CSVFlatDataParser(
-                List<String> columns,
-                String keyColumn
-        )
-        {
-            this(columns, keyColumn, false, 0, "");
-        }
-
-        //@JsonProperty
-        //public List<String> getColumns()
-        //{
-        //    return columns;
-        //}
-
-        @JsonProperty
-        public String getKeyColumn()
-        {
-            return this.keyColumn;
-        }
-
-        @Override
-        public Parser<String, List<String>> getParser()
-        {
-            return parser;
-        }
-
-        @Override
-        public boolean equals(final Object o)
-        {
-            if (this == o) {
-                return true;
-            }
-            if (o == null || getClass() != o.getClass()) {
-                return false;
-            }
-            final CSVFlatDataParser that = (CSVFlatDataParser) o;
-            return Objects.equals(columns, that.columns) &&
-                    Objects.equals(keyColumn, that.keyColumn);
-        }
-
-        @Override
-        public int hashCode()
-        {
-            return Objects.hash(columns, keyColumn);
-        }
-
-        @Override
-        public String toString()
-        {
-            return "CSVFlatDataParser{" +
-                    "columns=" + columns +
-                    ", keyColumn='" + keyColumn + '\'' +
-                    '}';
-        }
-    }
-
 
     @Override
     public String getLookupName() {
